requirements.txt
django = "~=2.2"
djangorestframework = "*"
pillow = "*"
coreapi = "*"
django-guardian = "*"
django-filter = "*"
django-debug-toolbar = "*"